Change the `mount_handlers` in INI to a more generic `data_sources`. Rename the `get_mount_handler` etc functions to `get_data_source`. Also add some stub methods to the `DataSource` class for incremental loading/ingestion.

+++ splitgraph/hooks/mount_handlers.py
@@ -1,41 +1,9 @@
 """
-Hooks for additional handlers used to mount other databases via FDW. These handlers become available
-in the command line tool (via `sgr mount`) and in the Splitfile interpreter (via `FROM MOUNT`).
+Extra wrapper code for mount handlers
 """
-import logging
-import types
-from importlib import import_module
-from typing import Dict, List, TYPE_CHECKING, Any, Type
+from typing import Dict, Any
 
-from splitgraph.config import CONFIG
-from splitgraph.config.config import get_all_in_section, get_singleton
-from splitgraph.config.keys import DEFAULTS
-from splitgraph.exceptions import MountHandlerError
-from splitgraph.hooks.data_source import ForeignDataWrapperDataSource
-
-if TYPE_CHECKING:
-    pass
-
-_MOUNT_HANDLERS: Dict[str, Type[ForeignDataWrapperDataSource]] = {}
-
-
-def get_mount_handler(mount_handler: str) -> Type[ForeignDataWrapperDataSource]:
-    """Returns a mount class for a given handler."""
-    try:
-        return _MOUNT_HANDLERS[mount_handler]
-    except KeyError:
-        raise MountHandlerError("Mount handler %s not supported!" % mount_handler)
-
-
-def get_mount_handlers() -> List[str]:
-    """Returns the names of all registered mount handlers."""
-    return list(_MOUNT_HANDLERS.keys())
-
-
-def register_mount_handler(name: str, mount_class: Type[ForeignDataWrapperDataSource]) -> None:
-    """Returns a data source under a given name. See `get_mount_handler` for the mount handler spec."""
-    global _MOUNT_HANDLERS
-    _MOUNT_HANDLERS[name] = mount_class
+from splitgraph.exceptions import DataSourceError
 
 
 def mount_postgres(mountpoint, **kwargs) -> None:
@@ -70,11 +38,19 @@ def mount(
     :param handler_kwargs: Dictionary of options to pass to the mount handler.
     :param overwrite: Delete the foreign server if it already exists. Used by mount_postgres for data pulls.
     """
+    # Workaround for circular imports
     from splitgraph.engine import get_engine
     from psycopg2.sql import Identifier, SQL
+    from splitgraph.hooks.data_source import get_data_source
+    from splitgraph.hooks.data_source.fdw import ForeignDataWrapperDataSource
 
     engine = get_engine()
-    data_source = get_mount_handler(mount_handler)
+    data_source = get_data_source(mount_handler)
+
+    if not data_source.supports_mount:
+        raise DataSourceError("Data source %s does not support mounting!")
+
+    assert issubclass(data_source, ForeignDataWrapperDataSource)
 
     engine.run_sql(SQL("DROP SCHEMA IF EXISTS {} CASCADE").format(Identifier(mountpoint)))
     engine.run_sql(
@@ -84,72 +60,3 @@ def mount(
     source = data_source.from_commandline(engine, handler_kwargs)
     source.mount(schema=mountpoint, overwrite=overwrite)
     engine.commit()
-
-
-def _register_default_handlers() -> None:
-    # Register the mount handlers from the config.
-    for handler_name, handler_class_name in get_all_in_section(CONFIG, "mount_handlers").items():
-        assert isinstance(handler_class_name, str)
-
-        try:
-            handler_class = _load_handler(handler_name, handler_class_name)
-
-            assert issubclass(handler_class, ForeignDataWrapperDataSource)
-            register_mount_handler(handler_name.lower(), handler_class)
-        except (ImportError, AttributeError) as e:
-            raise MountHandlerError(
-                "Error loading custom mount handler {0}".format(handler_name)
-            ) from e
-
-
-def _load_handler(handler_name, handler_class_name):
-    # Hack for old-style mount handlers that have now been moved -- don't crash and instead
-    # replace them in the config on the fly
-    handler_defaults = get_all_in_section(DEFAULTS, "mount_handlers")
-
-    fallback_used = False
-
-    try:
-        ix = handler_class_name.rindex(".")
-        handler_class = getattr(
-            import_module(handler_class_name[:ix]), handler_class_name[ix + 1 :]
-        )
-    except (ImportError, AttributeError):
-        if handler_name not in handler_defaults:
-            raise
-        handler_class_name = handler_defaults[handler_name]
-        ix = handler_class_name.rindex(".")
-        handler_class = getattr(
-            import_module(handler_class_name[:ix]), handler_class_name[ix + 1 :]
-        )
-        fallback_used = True
-
-    if isinstance(handler_class, types.FunctionType):
-        if handler_name not in handler_defaults:
-            raise MountHandlerError(
-                "Handler %s uses the old-style function interface which is not"
-                " compatible with this version of Splitgraph. "
-                "Delete it from your .sgconfig's [mount_handlers] section (%s)"
-                % (handler_name, get_singleton(CONFIG, "SG_CONFIG_FILE"))
-            )
-        handler_class_name = handler_defaults[handler_name]
-        ix = handler_class_name.rindex(".")
-        handler_class = getattr(
-            import_module(handler_class_name[:ix]), handler_class_name[ix + 1 :]
-        )
-        fallback_used = True
-    if fallback_used:
-        logging.warning(
-            "Handler %s uses the old-style function interface and was automatically replaced.",
-            handler_name,
-        )
-        logging.warning(
-            "Replace it with %s=%s in your .sgconfig's [mount_handlers] section (%s)",
-            handler_name,
-            handler_class_name,
-            get_singleton(CONFIG, "SG_CONFIG_FILE"),
-        )
-    return handler_class
-
-
-_register_default_handlers()
